requirements.txt
asgiref==3.2.7
astroid==2.4.1
beautifulsoup4==4.9.0
colorama==0.4.3
et-xmlfile==1.0.1
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
MouseInfo==0.1.3
numpy==1.18.3
openpyxl==3.0.3
Pillow==7.1.1
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
pylint==2.5.2
PyMsgBox==1.0.8
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
pytesseract==0.3.4
PyTweening==1.0.3
pytz==2020.1
selenium==3.141.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
toml==0.10.0
urllib3==1.25.9
wrapt==1.12.1
